Fold the shared reconcile algorithm into the durable driver
The two-driver seam (reconcile.ts with OpenStateSession, EmitFromStep and
the recoverFrom recovery paths) outlived its second driver: the in-process
reconciler was deleted when the runtime cut over, leaving the durable
driver as the sole caller and the recovery machinery unreachable. Inline
the algorithm into durable/operations.ts, preserving every step key, and
drop the dead recovery code. The RevConflict tests still expect conflicts
to propagate, which is what the durable driver always did.

Also extract withRuntime so deployApp, destroyApp and planApp share one
runtime acquire/close lifecycle instead of three copies.
