+++ packages/core/src/provisioner/workflows/workflow.plan.ts
@@ -1,6 +1,6 @@
 import { createPlan, type Plan } from "@notation/reconciler";
 import { getResourceGraph } from "src/orchestrator/graph";
-import { NodeDurableRuntime, resolveDeploymentId } from "../durable-runtime";
+import { withRuntime, type NodeDurableRuntime } from "../durable-runtime";
 
 export type { Plan, PlanNode, PlanDecision } from "@notation/reconciler";
 
@@ -20,19 +20,16 @@ export async function planApp({
   databasePath,
 }: PlanAppOptions): Promise<Plan> {
   const graph = await getResourceGraph(entryPoint);
-  const deploymentId =
-    suppliedRuntime?.deploymentId ?? resolveDeploymentId(entryPoint);
-  const runtime =
-    suppliedRuntime ?? new NodeDurableRuntime({ deploymentId, databasePath });
-  try {
-    await runtime.initialize();
-    return await createPlan({
-      resources: graph.resources,
-      state: runtime.state,
-      driftDetection,
-      maxOperationAttempts,
-    });
-  } finally {
-    if (!suppliedRuntime) runtime.close();
-  }
+  return withRuntime(
+    { entryPoint, runtime: suppliedRuntime, databasePath },
+    async (runtime) => {
+      await runtime.initialize();
+      return createPlan({
+        resources: graph.resources,
+        state: runtime.state,
+        driftDetection,
+        maxOperationAttempts,
+      });
+    },
+  );
 }
